Exp_DataProcessing/requirements.txt
psycopg2-binary==2.9.9
boto3==1.34.0
pandas==2.1.4
numpy==1.26.2

